Add LessonSubscription entity and update related models/classes
This commit includes the creation of the new entity, LessonSubscription, and several updates related to it. The new entity represents the many-to-many relationship between lessons and subscriptions. Updates include changes to the Lesson and Subscription entities to relate via LessonSubscription, adjustment in the Auto Mapper to handle this change, and a new error message for when students lack subscriptions. The IdentityService has also been updated to handle null user emails.